+++ GeosCore/diagnostics_mod.F90
@@ -1436,9 +1436,9 @@ CONTAINS
        ! SatDiagnOH: OH concentration [molec/cm3]:
        !---------------------------------------------------------------------
        IF ( State_Diag%Archive_SatDiagnOH ) THEN
-          State_Diag%SatDiagnOH(I,:,1:State_Grid%MaxChemLev) =               &
-               ( Spc(id_OH)%Conc(I,:,1:State_Grid%MaxChemLev) * good  *      &
-               State_Met%AIRDEN(I,:,1:State_Grid%MaxChemLev)  * AVO ) /      &
+          State_Diag%SatDiagnOH(I,:,1:State_Met%MaxChemLev) =               &
+               ( Spc(id_OH)%Conc(I,:,1:State_Met%MaxChemLev) * good  *      &
+               State_Met%AIRDEN(I,:,1:State_Met%MaxChemLev)  * AVO ) /      &
                ( State_Chm%SpcData(id_OH)%Info%MW_g ) / 1.0e+3_fp
        ENDIF