+++ GeosCore/gc_environment_mod.F90
@@ -386,8 +386,8 @@ CONTAINS
 !\\
 ! !INTERFACE:
 !
-  SUBROUTINE GC_Init_Extra( Diag_List,  Input_Opt,  State_Chm, &
-                            State_Diag, State_Grid, RC )
+  SUBROUTINE GC_Init_Extra( Diag_List,  Input_Opt,  State_Chm,               &
+                            State_Diag, State_Grid, State_Met, RC           )
 !
 ! !USES:
 !
@@ -415,6 +415,7 @@ CONTAINS
     USE State_Chm_Mod,      ONLY : ChmState
     USE State_Diag_Mod,     ONLY : DgnState
     USE State_Grid_Mod,     ONLY : GrdState
+    USE State_Met_Mod,      ONLY : MetState
     USE Sulfate_Mod,        ONLY : Init_Sulfate
     USE Tagged_CO_Mod,      ONLY : Init_Tagged_CO
     USE Tagged_O3_Mod,      ONLY : Init_Tagged_O3
@@ -431,6 +432,7 @@ CONTAINS
     TYPE(OptInput), INTENT(INOUT) :: Input_Opt   ! Input Options object
     TYPE(ChmState), INTENT(INOUT) :: State_Chm   ! Chemistry state object
     TYPE(DgnState), INTENT(INOUT) :: State_Diag  ! Diagnostics State object
+    TYPE(MetState), INTENT(INOUT) :: State_Met   ! Meteorology State object
 !
 ! !INPUT/OUTPUT PARAMETERS:
 !
@@ -483,7 +485,7 @@ CONTAINS
     !=======================================================================
     ! Initialize the hybrid pressure module.  Define Ap and Bp.
     !=======================================================================
-    CALL Init_Pressure( Input_Opt, State_Grid, RC )
+    CALL Init_Pressure( Input_Opt, State_Grid, State_Met, RC )
     IF ( RC /= GC_SUCCESS ) THEN
        ErrMsg = 'Error encountered in "Init_Pressure"!'
        CALL GC_Error( ErrMsg, RC, ThisLoc )
@@ -610,7 +612,8 @@ CONTAINS
     ! Initialize "sulfate_mod.F90"
     !-----------------------------------------------------------------
     IF ( Input_Opt%LSULF ) THEN
-       CALL Init_Sulfate( Input_Opt, State_Chm, State_Diag, State_Grid, RC )
+       CALL Init_Sulfate( Input_Opt,  State_Chm, State_Diag,                 &
+                          State_Grid, State_Met, RC                         )
        IF ( RC /= GC_SUCCESS ) THEN
           ErrMsg = 'Error encountered in "Init_Sulfate"!'
           CALL GC_Error( ErrMsg, RC, ThisLoc )
@@ -638,8 +641,8 @@ CONTAINS
     ! Carbon gases via KPP
     !-----------------------------------------------------------------
     IF ( Input_Opt%ITS_A_CARBON_SIM ) THEN
-       CALL Init_Carbon_Gases( Input_Opt,  State_Chm, State_Diag,             &
-                               State_Grid, RC                                )
+       CALL Init_Carbon_Gases( Input_Opt,  State_Chm, State_Diag,            &
+                               State_Grid, RC                               )
        IF ( RC /= GC_SUCCESS ) THEN
           ErrMsg = 'Error encountered in "Init_Carbon_Gases"!'
           CALL GC_Error( ErrMsg, RC, ThisLoc )
@@ -653,7 +656,8 @@ CONTAINS
     IF ( Input_Opt%ITS_A_MERCURY_SIM ) THEN
 
        ! Main mercury module
-       CALL Init_Mercury( Input_Opt, State_Grid, State_Chm, State_Diag, RC )
+       CALL Init_Mercury( Input_Opt,  State_Grid, State_Chm,                 &
+                          State_Diag, State_Met,  RC                        )
        IF ( RC /= GC_SUCCESS ) THEN
           ErrMsg = 'Error encountered in "Init_Mercury"!'
           CALL GC_Error( ErrMsg, RC, ThisLoc )