+++ GeosCore/ocean_mercury_mod.F90
@@ -378,7 +378,7 @@ CONTAINS
     ENDIF
 
     ! convert REAL*4 to REAL(fpp)
-    SO4_GC(:,:,1:State_Grid%MaxChemLev) = ARRAYso4(:,:,1:State_Grid%MaxChemLev)
+    SO4_GC(:,:,1:State_Met%MaxChemLev) = ARRAYso4(:,:,1:State_Met%MaxChemLev)
 
     !---------------------------
     ! Read NH4 from HEMCO
@@ -389,7 +389,7 @@ CONTAINS
     ENDIF
 
     ! convert REAL*4 to REAL(fpp)
-    NH4_CONC(:,:,1:State_Grid%MaxChemLev) = ARRAYnh4(:,:,1:State_Grid%MaxChemLev)
+    NH4_CONC(:,:,1:State_Met%MaxChemLev) = ARRAYnh4(:,:,1:State_Met%MaxChemLev)
 
     !---------------------------
     ! Read NIT from HEMCO
@@ -400,7 +400,7 @@ CONTAINS
     ENDIF
 
     ! convert REAL*4 to REAL(fpp)
-    NIT_CONC(:,:,1:State_Grid%MaxChemLev) = ARRAYnit(:,:,1:State_Grid%MaxChemLev)
+    NIT_CONC(:,:,1:State_Met%MaxChemLev) = ARRAYnit(:,:,1:State_Met%MaxChemLev)
 
     !---------------------------
     ! Read BCPI+BCPO from HEMCO
@@ -425,7 +425,7 @@ CONTAINS
     ! convert REAL*4 to REAL(fpp)
     ARRAYtemp           = 0.e0_f4
     ARRAYtemp(:,:,1:LL) = ARRAYbcpi + ARRAYbcpo
-    BC_CONC(:,:,1:State_Grid%MaxChemLev) = ARRAYtemp(:,:,1:State_Grid%MaxChemLev)
+    BC_CONC(:,:,1:State_Met%MaxChemLev) = ARRAYtemp(:,:,1:State_Met%MaxChemLev)
 
     !---------------------------
     ! Read OCPI+OCPO from HEMCO
@@ -450,7 +450,7 @@ CONTAINS
     ! convert REAL*4 to REAL(fpp)
     ARRAYtemp           = 0.e0_f4
     ARRAYtemp(:,:,1:LL) = ARRAYocpi + ARRAYocpo
-    OC_CONC(:,:,1:State_Grid%MaxChemLev) = ARRAYtemp(:,:,1:State_Grid%MaxChemLev)
+    OC_CONC(:,:,1:State_Met%MaxChemLev) = ARRAYtemp(:,:,1:State_Met%MaxChemLev)
 
     !---------------------------
     ! Read DST1 from HEMCO
@@ -461,7 +461,7 @@ CONTAINS
     ENDIF
 
     ! convert REAL*4 to REAL(fpp)
-    DST_CONC(:,:,1:State_Grid%MaxChemLev) = ARRAYdst1(:,:,1:State_Grid%MaxChemLev)
+    DST_CONC(:,:,1:State_Met%MaxChemLev) = ARRAYdst1(:,:,1:State_Met%MaxChemLev)
 
     !$OMP PARALLEL DO                                                        &
     !$OMP DEFAULT( SHARED                                                   )&